+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-## 2.31.2
+## 2.32.0
+### Features
+* add insert typing animation
 ### Bugfix
 * fix calculating number of rows (affecting less command)
+* fix glow with prism and error messages [#729](https://github.com/jcubic/jquery.terminal/issues/729)
+* fix prism and typing animation [#726](https://github.com/jcubic/jquery.terminal/issues/726)
+* fix various typing errors
 
 ## 2.31.1
 ### Bugfix